black76/requirements.txt
Django==3.1.14
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.11.2
factory-boy==2.12.0
pytest==5.4.3
pytest-django==3.9.0